+++ packages/web3-eth2-beaconchain/test/axiosMock.ts
@@ -1,0 +1,3 @@
+import jest from 'jest'
+
+export const get = jest.fn(() => Promise.resolve({ data: {} }))