+++ CHANGELOG.md
@@ -300,3 +300,4 @@ Released with 1.0.0-beta.37 code base.
 ### Fixed
 
 - Fixed decoding bytes and string parameters for logs emitted with solc 0.4.x (#3724, #3738)
+- Grammar changes to inputAddressFormatter error message