+++ CHANGELOG.md
@@ -291,6 +291,7 @@ Released with 1.0.0-beta.37 code base.
 ## [1.3.1]
 
 ### Changed
+
 - bump utils 0.10.0^ -> 0.12.0 (#3733)
 
 ### Removed
@@ -299,5 +300,6 @@ Released with 1.0.0-beta.37 code base.
 
 ### Fixed
 
+- Fix possible unhandled promise rejection when sending a transaction (#3708)
 - Fixed decoding bytes and string parameters for logs emitted with solc 0.4.x (#3724, #3738)
 - Grammar changes to inputAddressFormatter error message