+++ CHANGELOG.md
@@ -294,6 +294,7 @@ Released with 1.0.0-beta.37 code base.
 
 - Add `web3-eth2-core` package (#3743)
 - Add `web3-eth2-beaconchain` package (#3743)
+- Add `stripHexPrefix` method to `web3-utils` package (#3776)
 
 ### Changed